refactor(ui): move plan-reminder form seeds into helpers
PlanReminderFormSeed and the create/template/edit/duplicate seed
builders are pure mappings — relocate them next to the other form
helpers so PlanReminderFormDialog stays the React owner of field state
and submit, matching the #1044 ownership split.
