+++ packages/ui/src/plan-reminder-helpers.ts
@@ -17,6 +17,7 @@
  */
 
 import type {
+  BotProvider,
   PlanReminder,
   PlanReminderDeliveryTarget,
   PlanReminderRecurrence,
@@ -296,4 +297,79 @@ export function runStatusLabel(status: NonNullable<PlanReminder['lastRun']>['sta
   if (status === 'triggered') return '已触发';
   if (status === 'blocked') return '已阻止';
   return '失败';
+}
+
+/**
+ * Initial field values for the plan-reminder form dialog
+ * (PlanReminderFormDialog, issue #1044). The panel builds one seed per open
+ * (create / template / edit / duplicate) and the dialog mounts with it, so
+ * the seeds are pure mappings — they live here with the other form helpers,
+ * not in the component file.
+ */
+export interface PlanReminderFormSeed {
+  editingId: string | null;
+  title: string;
+  note: string;
+  runAtLocal: string;
+  recurrence: PlanReminderRecurrence;
+  cronExpression: string;
+  deliveryChannel: PlanReminderDeliveryTarget['channel'];
+  deliveryPlatform: BotProvider;
+  deliveryChatId: string;
+}
+
+/** Blank create-mode seed (one hour from now, no recurrence, local delivery). */
+export function createPlanReminderFormSeed(): PlanReminderFormSeed {
+  return {
+    editingId: null,
+    title: '',
+    note: '',
+    runAtLocal: toPlanReminderDateTimeInputValue(Date.now() + 60 * 60 * 1000),
+    recurrence: 'none',
+    cronExpression: '0 9 * * 1-5',
+    deliveryChannel: 'local',
+    deliveryPlatform: 'telegram',
+    deliveryChatId: '',
+  };
+}
+
+/** Create-mode seed prefilled from an example template. */
+export function planReminderTemplateSeed(template: PlanReminderExampleTemplate): PlanReminderFormSeed {
+  return {
+    ...createPlanReminderFormSeed(),
+    title: template.title,
+    note: template.note,
+    recurrence: template.recurrence,
+    cronExpression: template.cronExpression,
+    runAtLocal: toPlanReminderDateTimeInputValue(planReminderTemplateNextRunAt(template)),
+  };
+}
+
+function planReminderReminderSeed(reminder: PlanReminder): PlanReminderFormSeed {
+  return {
+    editingId: reminder.id,
+    title: reminder.title,
+    note: reminder.note,
+    runAtLocal: toPlanReminderDateTimeInputValue(planReminderEditableRunAt(reminder)),
+    recurrence: planReminderRecurrenceValue(reminder),
+    cronExpression: reminder.schedule.kind === 'cron' ? reminder.schedule.expression : '0 9 * * 1-5',
+    deliveryChannel: reminder.delivery.channel,
+    ...(reminder.delivery.channel === 'bot'
+      ? { deliveryPlatform: reminder.delivery.platform, deliveryChatId: reminder.delivery.chatId }
+      : { deliveryPlatform: 'telegram' as BotProvider, deliveryChatId: '' }),
+  };
+}
+
+/** Edit-mode seed prefilled from an existing reminder. */
+export function planReminderEditSeed(reminder: PlanReminder): PlanReminderFormSeed {
+  return planReminderReminderSeed(reminder);
+}
+
+/** Create-mode seed copying an existing reminder under a 副本 title. */
+export function planReminderDuplicateSeed(reminder: PlanReminder): PlanReminderFormSeed {
+  return {
+    ...planReminderReminderSeed(reminder),
+    editingId: null,
+    title: duplicatePlanReminderTitle(reminder.title),
+  };
 }
