+++ packages/ui/src/plan-reminder-panel.tsx
@@ -25,24 +25,22 @@ import {
 import {
   PLAN_REMINDER_EXAMPLE_TEMPLATES,
   type PlanReminderExampleTemplate,
+  type PlanReminderFormSeed,
   comparePlanReminderBySort,
+  createPlanReminderFormSeed,
   formatPlanRecurrence,
   formatReminderCountdown,
   formatReminderTime,
   normalizePlanReminderSearchQuery,
+  planReminderDuplicateSeed,
+  planReminderEditSeed,
   planReminderMatchesSearch,
   planReminderRunRangeStart,
   planReminderStatusLabel,
+  planReminderTemplateSeed,
   runStatusLabel,
 } from './plan-reminder-helpers.js';
-import {
-  PlanReminderFormDialog,
-  createPlanReminderFormSeed,
-  planReminderDuplicateSeed,
-  planReminderEditSeed,
-  planReminderTemplateSeed,
-  type PlanReminderFormSeed,
-} from './plan-reminder-form-dialog.js';
+import { PlanReminderFormDialog } from './plan-reminder-form-dialog.js';
 import { PlanReminderSelect } from './plan-reminder-select.js';
 import {
   Button as UiButton,
